+++ CHANGELOG.md
@@ -1,15 +1,41 @@
 ## CHANGELOG
 
 lists items that might need to run manually.
+
+* 20250612
+  The package provides the following listener:
+
+* **`ActivityLogMessageSentListener`**: Listen to the MessageSent event of the Mailable to create the corresponding
+  communication log and activity log
+* must also register the event in the **`EventServiceProvider`**
+
+```php
+
+use Dcodegroup\ActivityLog\Listeners\ActivityLogMessageSentListener;
+use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+use Illuminate\Mail\Events\MessageSent;
+
+
+class EventServiceProvider extends ServiceProvider
+{
+    protected $listen = [
+
+        MessageSent::class => [
+            ActivityLogMessageSentListener::class,
+        ],
+    ];
+}
+
+```
+
 * 20250509
 
 Added – Support for `extra_models` parameter in the Activity Log API endpoint.
 `extra_models`: A comma-separated list of relationship method names to include related models' activities.
 
-When provided, the system will dynamically load specified relationships, and include related models' activities in the query using `orWhere` clauses.
+When provided, the system will dynamically load specified relationships, and include related models' activities in the
+query using `orWhere` clauses.
 This enhancement allows for broader and more flexible activity retrieval across related models.
-
-
 
 * 20250502